requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
scapy>=2.5
tensorflow>=2.16
liac-arff>=2.5.0
jupyter>=1.0
ipykernel>=6.0
